Fix 32 clippy errors for CI release gate
- approval.rs: Add #[derive(Default)], remove unnecessary usize casts
- context_manager.rs: Remove unnecessary u64 cast
- tool_output.rs: Remove unnecessary let binding
- interactive.rs: Collapse if statements, remove unnecessary let binding
- instructions.rs: Collapse if statements, use is_some_and()
- slash_commands.rs: Collapse if statements
- agent_loop.rs: Collapse if statements, remove needless borrow
- file_search.rs: Collapse if statements, use io::Error::other()
- state_manager.rs: Use io::Error::other()
- input.rs: Fix doc comment indentation, remove needless reference
- edit_file.rs: Collapse 5 nested if statements in rollback loops
- rename_symbol.rs: Collapse nested if statements
- minimax.rs: Collapse nested if statements

Validation:
- cargo clippy --lib -- -D warnings: PASSED
- cargo test --lib: 980/980 passed

+++ src/core/context/context_manager.rs
@@ -152,7 +152,7 @@ pub fn get_new_context_messages_and_metadata(
             .unwrap_or((256_000.0 * threshold_pct) as u64);
 
         if total_tokens >= max_allowed_size {
-            let keep = if (total_tokens as u64 / 2) > max_allowed_size {
+            let keep = if (total_tokens / 2) > max_allowed_size {
                 TruncationKeep::LastQuarter
             } else {
                 TruncationKeep::Half
